Refactor the plot functions submit
